Refactor GameTracker.tsx into the directories, client/src/pages/GameTracker/..., utils, hooks, components
